// src/app/coinnovation/components/Types/CoinnovationTypes.tsx
export interface ProjectData {
  project_id: string;
  project_name: string;
  priority: string;
  status: string;
  start_date: string;
  end_date: string;
  group_company: string;
  enterprise_img: any;
  enterprise: string;
  owner: string;
  approver: string;
  category: string;
  department: string;
  business_unit: string;
  location: string;
  project_description: string;
  problem_statement: string;
  context: string;
}